Crate shiplift

Shiplift is a multi-transport utility for maneuvering docker containers

examples

use tokio::prelude::Future;

let docker = shiplift::Docker::new();
let fut = docker.images().list(&Default::default()).map(|images| {
  println!("docker images in stock");
  for i in images {
    println!("{:?}", i.repo_tags);
  }
}).map_err(|e| eprintln!("Something bad happened! {}", e));

tokio::run(fut);

Re-exports

pub use crate::builder::BuildOptions;

pub use crate::builder::ContainerConnectionOptions;

pub use crate::builder::ContainerFilter;

pub use crate::builder::ContainerListOptions;

pub use crate::builder::ContainerOptions;

pub use crate::builder::EventsOptions;

pub use crate::builder::ExecContainerOptions;

pub use crate::builder::ImageFilter;

pub use crate::builder::ImageListOptions;

pub use crate::builder::LogsOptions;

pub use crate::builder::NetworkCreateOptions;

pub use crate::builder::NetworkListOptions;

pub use crate::builder::PullOptions;

pub use crate::builder::RegistryAuth;

pub use crate::builder::RmContainerOptions;

pub use crate::builder::TagOptions;

pub use crate::builder::VolumeCreateOptions;

pub use crate::errors::Error;

Modules

builder
Interfaces for building various structures

errors
Representations of various client errors

read

rep
Rust representations of docker json structures

transport
Transports for communicating with the docker daemon

tty

Structs

Container
Interface for accessing and manipulating a docker container

Containers
Interface for docker containers

Docker
Entrypoint interface for communicating with docker daemon

Image
Interface for accessing and manipulating a named docker image

Images
Interface for docker images

Network
Interface for accessing and manipulating a docker network

Networks
Interface for docker network

Volume
Interface for accessing and manipulating a named docker volume

Volumes
Interface for docker volumes

Type Definitions

Result
Represents the result of all docker operations