Struct shiplift::builder::ContainerOptions

pub struct ContainerOptions {
    pub name: Option<String>,
    // some fields omitted
}

Interface for building a new docker container from an existing image

Fields
name: Option<String>
Methods

impl ContainerOptions

fn builder(name: &str) -> ContainerOptionsBuilder

return a new instance of a builder for options

fn serialize(&self) -> Result<String>

serialize options as a string. returns None if no options are defined

fn parse_from<'a, K, V>(&self,
params: &'a HashMap<K, V>,
host_config: &mut BTreeMap<String, Json>,
body: &mut BTreeMap<String, Json>) where &'a HashMap<K, V>: IntoIterator, K: ToString + Eq + Hash, V: ToJson