Struct shiplift::builder::ContainerOptionsBuilder

pub struct ContainerOptionsBuilder { /* fields omitted */ }

Implementations

impl ContainerOptionsBuilder

pub fn name(&mut self, name: &str) -> &mut Self

pub fn working_dir(&mut self, working_dir: &str) -> &mut Self

Specify the working dir (corresponds to the -w docker cli argument)

pub fn volumes(&mut self, volumes: Vec<&str>) -> &mut Self

Specify any bind mounts, taking the form of /some/host/path:/some/container/path

pub fn publish_all_ports(&mut self) -> &mut Self

enable all exposed ports on the container to be mapped to random, available, ports on the host

pub fn expose(
&mut self,
srcport: u32,
protocol: &str,
hostport: u32
) -> &mut Self

pub fn publish(&mut self, srcport: u32, protocol: &str) -> &mut Self

Publish a port in the container without assigning a port on the host

pub fn links(&mut self, links: Vec<&str>) -> &mut Self

pub fn memory(&mut self, memory: u64) -> &mut Self

pub fn memory_swap(&mut self, memory_swap: i64) -> &mut Self

Total memory limit (memory + swap) in bytes. Set to -1 (default) to enable unlimited swap.

pub fn nano_cpus(&mut self, nano_cpus: u64) -> &mut Self

CPU quota in units of 10-9 CPUs. Set to 0 (default) for there to be no limit.

For example, setting nano_cpus to 500_000_000 results in the container being allocated 50% of a single CPU, while 2_000_000_000 results in the container being allocated 2 CPUs.

pub fn cpus(&mut self, cpus: f64) -> &mut Self

CPU quota in units of CPUs. This is a wrapper around nano_cpus to do the unit conversion.

See nano_cpus.

pub fn cpu_shares(&mut self, cpu_shares: u32) -> &mut Self

Sets an integer value representing the container's relative CPU weight versus other containers.

pub fn labels(&mut self, labels: &HashMap<&str, &str>) -> &mut Self

pub fn attach_stdin(&mut self, attach: bool) -> &mut Self

Whether to attach to stdin.

pub fn attach_stdout(&mut self, attach: bool) -> &mut Self

Whether to attach to stdout.

pub fn attach_stderr(&mut self, attach: bool) -> &mut Self

Whether to attach to stderr.

pub fn tty(&mut self, tty: bool) -> &mut Self

Whether standard streams should be attached to a TTY.

pub fn extra_hosts(&mut self, hosts: Vec<&str>) -> &mut Self

pub fn volumes_from(&mut self, volumes: Vec<&str>) -> &mut Self

pub fn network_mode(&mut self, network: &str) -> &mut Self

pub fn env<E, S>(&mut self, envs: E) -> &mut Self where
S: AsRef<str> + Serialize,
E: AsRef<[S]> + Serialize,

pub fn cmd(&mut self, cmds: Vec<&str>) -> &mut Self

pub fn entrypoint(&mut self, entrypoint: &str) -> &mut Self

pub fn capabilities(&mut self, capabilities: Vec<&str>) -> &mut Self

pub fn devices(&mut self, devices: Vec<HashMap<String, String>>) -> &mut Self

pub fn log_driver(&mut self, log_driver: &str) -> &mut Self

pub fn restart_policy(
&mut self,
name: &str,
maximum_retry_count: u64
) -> &mut Self

pub fn auto_remove(&mut self, set: bool) -> &mut Self

pub fn stop_signal(&mut self, sig: &str) -> &mut Self

Signal to stop a container as a string. Default is "SIGTERM".

pub fn stop_signal_num(&mut self, sig: u64) -> &mut Self

Signal to stop a container as an integer. Default is 15 (SIGTERM).

pub fn stop_timeout(&mut self, timeout: Duration) -> &mut Self

Timeout to stop a container. Only seconds are counted. Default is 10s

pub fn userns_mode(&mut self, mode: &str) -> &mut Self

pub fn privileged(&mut self, set: bool) -> &mut Self

pub fn user(&mut self, user: &str) -> &mut Self

pub fn build(&self) -> ContainerOptions