Struct shiplift::network::Networks

pub struct Networks<'docker> { /* fields omitted */ }

Interface for docker network

API Reference: https://docs.docker.com/engine/api/v1.41/#tag/Network

Implementations

impl<'docker> Networks<'docker>

pub fn new(docker: &'docker Docker) -> Self

Exports an interface for interacting with docker Networks

pub async fn list(&self, opts: &NetworkListOptions) -> Result<Vec<NetworkInfo>>

List the docker networks on the current docker host

API Reference: https://docs.docker.com/engine/api/v1.41/#operation/NetworkList

pub fn get<S>(&self, id: S) -> Network<'docker> where
S: Into<String>,

Returns a reference to a set of operations available to a specific network instance

pub async fn create(
&self,
opts: &NetworkCreateOptions
) -> Result<NetworkCreateInfo>

Create a new Network instance

API Reference: https://docs.docker.com/engine/api/v1.41/#operation/NetworkCreate