Struct shiplift::Container

pub struct Container<'a, 'b> {
    // some fields omitted
}

Interface for accessing and manipulating a docker container

Methods

impl<'a, 'b> Container<'a, 'b>

fn new(docker: &'a Docker, id: &'b str) -> Container<'a, 'b>

Exports an interface exposing operations against a container instance

fn inspect(&self) -> Result<ContainerDetails>

Inspects the current docker container instance's details

fn top(&self, psargs: Option<&str>) -> Result<Top>

Returns a top view of information about the container process

fn logs(&self, opts: &LogsOptions) -> Result<Box<Read>>

Returns a stream of logs emitted but the container instance

fn changes(&self) -> Result<Vec<Change>>

Returns a set of changes made to the container instance

fn export(&self) -> Result<Box<Read>>

Exports the current docker container into a tarball

fn stats(&self) -> Result<Box<Iterator<Item=Stats>>>

Returns a stream of stats specific to this container instance

fn start(&'a self) -> Result<()>

Start the container instance

fn stop(&self, wait: Option<Duration>) -> Result<()>

Stop the container instance

fn restart(&self, wait: Option<Duration>) -> Result<()>

Restart the container instance

fn kill(&self, signal: Option<&str>) -> Result<()>

Kill the container instance

fn rename(&self, name: &str) -> Result<()>

Rename the container instance

fn pause(&self) -> Result<()>

Pause the container instance

fn unpause(&self) -> Result<()>

Unpause the container instance

fn wait(&self) -> Result<Exit>

Wait until the container stops

fn delete(&self) -> Result<()>

Delete the container instance (todo: force/v)