Struct shiplift::Docker

pub struct Docker { /* fields omitted */ }

Entrypoint interface for communicating with docker daemon

Methods

impl Docker

fn new() -> Docker

constructs a new Docker instance for a docker host listening at a url specified by an env var DOCKER_HOST, falling back on unix:///var/run/docker.sock

fn host(host: Url) -> Docker

constructs a new Docker instance for docker host listening at the given host url

fn images<'a>(&'a self) -> Images

Exports an interface for interacting with docker images

fn containers<'a>(&'a self) -> Containers

Exports an interface for interacting with docker containers

fn version(&self) -> Result<Version>

Returns version information associated with the docker daemon

fn info(&self) -> Result<Info>

Returns information associated with the docker daemon

fn ping(&self) -> Result<String>

Returns a simple ping response indicating the docker daemon is accessible

fn events(&self, opts: &EventsOptions) -> Result<Box<Iterator<Item=Event>>>

Returns an interator over streamed docker events