Struct shiplift::Image

pub struct Image<'a, 'b> { /* fields omitted */ }

Interface for accessing and manipulating a named docker image

Methods

impl<'a, 'b> Image<'a, 'b>

fn new<S>(docker: &'a Docker, name: S) -> Image<'a, 'b> where S: Into<Cow<'b, str>>

Exports an interface for operations that may be performed against a named image

fn inspect(&self) -> Result<ImageDetails>

Inspects a named image's details

fn history(&self) -> Result<Vec<History>>

Lists the history of the images set of changes

fn delete(&self) -> Result<Vec<Status>>

Delete's an image

fn export(&self) -> Result<Box<Read>>

Export this image to a tarball