Struct shiplift::Images

pub struct Images<'a> {
    // some fields omitted
}

Interface for docker images

Methods

impl<'a> Images<'a>

fn new(docker: &'a Docker) -> Images<'a>

Exports an interface for interacting with docker images

fn build(&self, opts: &BuildOptions) -> Result<Box<Iterator<Item=BuildOutput>>>

Builds a new image build by reading a Dockerfile in a target directory

fn list(&self, opts: &ImageListOptions) -> Result<Vec<ImageRep>>

Lists the docker images on the current docker host

fn get(&'a self, name: &'a str) -> Image

Returns a reference to a set of operations available for a named image

fn search(&self, term: &str) -> Result<Vec<SearchResult>>

Search for docker images by term

fn pull(&self, opts: &PullOptions) -> Result<Box<Iterator<Item=PullOutput>>>

Pull and create a new docker images from an existing image

fn export(&self, names: Vec<&str>) -> Result<Box<Read>>

exports a collection of named images, either by name, name:tag, or image id, into a tarball