Struct shiplift::Networks

pub struct Networks<'a> { /* fields omitted */ }

Interface for docker network

Methods

impl<'a> Networks<'a>

pub fn new(docker: &'a Docker) -> Networks<'a>

Exports an interface for interacting with docker Networks

pub fn list(&self, opts: &NetworkListOptions) -> Result<Vec<NetworkInfo>>

List the docker networks on the current docker host

pub fn get(&'a self, id: &'a str) -> Network

Returns a reference to a set of operations available to a specific network instance

pub fn create(
&'a self,
opts: &NetworkCreateOptions
) -> Result<NetworkCreateInfo>