Enum shiplift::transport::Transport

pub enum Transport {
    Tcp {
        client: Client<HttpConnector>,
        host: String,
    },
    EncryptedTcp {
        client: Client<HttpsConnector<HttpConnector>>,
        host: String,
    },
    Unix {
        client: Client<UnixConnector>,
        path: String,
    },
}

Transports are types which define the means of communication with the docker daemon

Variants

Tcp

A network tcp interface

Fields of Tcp

client: Client<HttpConnector>host: String

EncryptedTcp

TCP/TLS

Fields of EncryptedTcp

client: Client<HttpsConnector<HttpConnector>>host: String

Unix

A Unix domain socket

Fields of Unix

client: Client<UnixConnector>path: String

Methods

impl Transport

pub fn request<B>(
&self,
method: Method,
endpoint: &str,
body: Option<(B, Mime)>
) -> impl Future<Item = String, Error = Error> where
B: Into<Body>,

Make a request and return the whole response in a String

pub fn stream_chunks<B, H>(
&self,
method: Method,
endpoint: &str,
body: Option<(B, Mime)>,
headers: Option<H>
) -> impl Stream<Item = Chunk, Error = Error> where
B: Into<Body>,
H: IntoIterator<Item = (&'static str, String)>,

Make a request and return a Stream of Chunks as they are returned.

pub fn stream_upgrade<B>(
&self,
method: Method,
endpoint: &str,
body: Option<(B, Mime)>
) -> impl Future<Item = impl AsyncRead + AsyncWrite, Error = Error> where
B: Into<Body>,

Makes an HTTP request, upgrading the connection to a TCP stream on success.

This method can be used for operations such as viewing docker container logs interactively.

pub fn stream_upgrade_multiplexed<B>(
&self,
method: Method,
endpoint: &str,
body: Option<(B, Mime)>
) -> impl Future<Item = Multiplexed, Error = Error> where
B: Into<Body> + 'static,