Module shiplift::volume

Create and manage persistent storage that can be attached to containers.

API Reference: https://docs.docker.com/engine/api/v1.41/#tag/Volume

Structs

Volume
Interface for accessing and manipulating a named docker volume

VolumeCreateInfo

VolumeCreateOptions
Interface for creating volumes

VolumeCreateOptionsBuilder

VolumeInfo

Volumes
Interface for docker volumes

VolumesInfo