Crate shiplift

Shiplift is a multi-transport utility for maneuvering docker containers

examples

let docker = shiplift::Docker::new();

match docker.images().list(&Default::default()).await {
    Ok(images) => {
        for image in images {
            println!("{:?}", image.repo_tags);
        }
    },
    Err(e) => eprintln!("Something bad happened! {}", e),
}

Re-exports

pub use crate::builder::BuildOptions;

pub use crate::builder::ContainerConnectionOptions;

pub use crate::builder::ContainerFilter;

pub use crate::builder::ContainerListOptions;

pub use crate::builder::ContainerOptions;

pub use crate::builder::EventsOptions;

pub use crate::builder::ExecContainerOptions;

pub use crate::builder::ExecResizeOptions;

pub use crate::builder::ImageFilter;

pub use crate::builder::ImageListOptions;

pub use crate::builder::LogsOptions;

pub use crate::builder::NetworkCreateOptions;

pub use crate::builder::NetworkListOptions;

pub use crate::builder::PullOptions;

pub use crate::builder::RegistryAuth;

pub use crate::builder::RmContainerOptions;

pub use crate::builder::TagOptions;

pub use crate::builder::VolumeCreateOptions;

pub use crate::errors::Error;

Modules

builder
Interfaces for building various structures

errors
Representations of various client errors

rep
Rust representations of docker json structures

transport
Transports for communicating with the docker daemon

tty
Types for working with docker TTY streams

Structs

Container
Interface for accessing and manipulating a docker container

Containers
Interface for docker containers

Docker
Entrypoint interface for communicating with docker daemon

Exec
Interface for docker exec instance

Image
Interface for accessing and manipulating a named docker image

Images
Interface for docker images

Network
Interface for accessing and manipulating a docker network

Networks
Interface for docker network

Uri
The URI component of a request.

Volume
Interface for accessing and manipulating a named docker volume

Volumes
Interface for docker volumes

Type Definitions

Result
Represents the result of all docker operations