Struct shiplift::builder::BuildOptionsBuilder

pub struct BuildOptionsBuilder { /* fields omitted */ }

Methods

impl BuildOptionsBuilder

pub fn new<S>(path: S) -> BuildOptionsBuilder where
S: Into<String>,

path is expected to be a file path to a directory containing a Dockerfile describing how to build a Docker image

pub fn dockerfile<P>(&mut self, path: P) -> &mut BuildOptionsBuilder where
P: Into<String>,

set the name of the docker file. defaults to "DockerFile"

pub fn tag<T>(&mut self, t: T) -> &mut BuildOptionsBuilder where
T: Into<String>,

tag this image with a name after building it

pub fn remote<R>(&mut self, r: R) -> &mut BuildOptionsBuilder where
R: Into<String>,

pub fn nocache<R>(&mut self, nc: bool) -> &mut BuildOptionsBuilder

don't use the image cache when building image

pub fn rm(&mut self, r: bool) -> &mut BuildOptionsBuilder

pub fn forcerm(&mut self, fr: bool) -> &mut BuildOptionsBuilder

pub fn network_mode<T>(&mut self, t: T) -> &mut BuildOptionsBuilder where
T: Into<String>,

bridge, host, none, container:<name|id>, or a custom network name.

pub fn build(&self) -> BuildOptions