Struct shiplift::Container

pub struct Container<'a, 'b> { /* fields omitted */ }

Interface for accessing and manipulating a docker container

Methods

impl<'a, 'b> Container<'a, 'b>

pub fn new<S>(docker: &'a Docker, id: S) -> Container<'a, 'b> where
S: Into<Cow<'b, str>>,

Exports an interface exposing operations against a container instance

pub fn id(&self) -> &str

a getter for the container id

pub fn inspect(&self) -> Result<ContainerDetails>

Inspects the current docker container instance's details

pub fn top(&self, psargs: Option<&str>) -> Result<Top>

Returns a top view of information about the container process

pub fn logs(&self, opts: &LogsOptions) -> Result<Box<Read>>

Returns a stream of logs emitted but the container instance

pub fn changes(&self) -> Result<Vec<Change>>

Returns a set of changes made to the container instance

pub fn export(&self) -> Result<Box<Read>>

Exports the current docker container into a tarball

pub fn stats(&self) -> Result<Box<Iterator<Item = Stats>>>

Returns a stream of stats specific to this container instance

pub fn start(&'a self) -> Result<()>

Start the container instance

pub fn stop(&self, wait: Option<Duration>) -> Result<()>

Stop the container instance

pub fn restart(&self, wait: Option<Duration>) -> Result<()>

Restart the container instance

pub fn kill(&self, signal: Option<&str>) -> Result<()>

Kill the container instance

pub fn rename(&self, name: &str) -> Result<()>

Rename the container instance

pub fn pause(&self) -> Result<()>

Pause the container instance

pub fn unpause(&self) -> Result<()>

Unpause the container instance

pub fn wait(&self) -> Result<Exit>

Wait until the container stops

pub fn delete(&self) -> Result<()>

Delete the container instance

Use remove instead to use the force/v options.

pub fn remove(&self, opts: RmContainerOptions) -> Result<()>

Delete the container instance (todo: force/v)

pub fn exec(&self, opts: &ExecContainerOptions) -> Result<Tty>

Exec the specified command in the container