Struct shiplift::Container

pub struct Container<'a, 'b> { /* fields omitted */ }

Interface for accessing and manipulating a docker container

Methods

impl<'a, 'b> Container<'a, 'b>

pub fn new<S>(docker: &'a Docker, id: S) -> Container<'a, 'b> where
S: Into<Cow<'b, str>>,

Exports an interface exposing operations against a container instance

pub fn id(&self) -> &str

a getter for the container id

pub fn inspect(
&self
) -> impl Future<Item = ContainerDetails, Error = Error>

Inspects the current docker container instance's details

pub fn top(
&self,
psargs: Option<&str>
) -> impl Future<Item = Top, Error = Error>

Returns a top view of information about the container process

pub fn logs(
&self,
opts: &LogsOptions
) -> impl Stream<Item = Chunk, Error = Error>

Returns a stream of logs emitted but the container instance

pub fn attach(
&self
) -> impl Future<Item = Multiplexed, Error = Error>

Attaches to a running container, returning a stream that can be used to interact with the standard IO streams.

pub fn attach_blocking(&self) -> Result<MultiplexedBlocking>

Attaches to a running container, returning a stream that can be used to interact with the standard IO streams.

pub fn changes(
&self
) -> impl Future<Item = Vec<Change>, Error = Error>

Returns a set of changes made to the container instance

pub fn export(
&self
) -> impl Stream<Item = Vec<u8>, Error = Error>

Exports the current docker container into a tarball

pub fn stats(
&self
) -> impl Stream<Item = Stats, Error = Error>

Returns a stream of stats specific to this container instance

pub fn start(
&self
) -> impl Future<Item = (), Error = Error>

Start the container instance

pub fn stop(
&self,
wait: Option<Duration>
) -> impl Future<Item = (), Error = Error>

Stop the container instance

pub fn restart(
&self,
wait: Option<Duration>
) -> impl Future<Item = (), Error = Error>

Restart the container instance

pub fn kill(
&self,
signal: Option<&str>
) -> impl Future<Item = (), Error = Error>

Kill the container instance

pub fn rename(
&self,
name: &str
) -> impl Future<Item = (), Error = Error>

Rename the container instance

pub fn pause(
&self
) -> impl Future<Item = (), Error = Error>

Pause the container instance

pub fn unpause(
&self
) -> impl Future<Item = (), Error = Error>

Unpause the container instance

pub fn wait(
&self
) -> impl Future<Item = Exit, Error = Error>

Wait until the container stops

pub fn delete(
&self
) -> impl Future<Item = (), Error = Error>

Delete the container instance

Use remove instead to use the force/v options.

pub fn remove(
&self,
opts: RmContainerOptions
) -> impl Future<Item = (), Error = Error>

Delete the container instance (todo: force/v)

pub fn exec(
&self,
opts: &ExecContainerOptions
) -> impl Stream<Item = Chunk, Error = Error>

Exec the specified command in the container