Struct shiplift::Containers

pub struct Containers<'a> { /* fields omitted */ }

Interface for docker containers

Methods

impl<'a> Containers<'a>

pub fn new(docker: &'a Docker) -> Containers<'a>

Exports an interface for interacting with docker containers

pub fn list(&self, opts: &ContainerListOptions) -> Result<Vec<ContainerRep>>

Lists the container instances on the docker host

pub fn get(&'a self, name: &'a str) -> Container

Returns a reference to a set of operations available to a specific container instance

pub fn create(&'a self, opts: &ContainerOptions) -> Result<ContainerCreateInfo>

Returns a builder interface for creating a new container instance