Struct shiplift::Docker

pub struct Docker { /* fields omitted */ }

Entrypoint interface for communicating with docker daemon

Methods

impl Docker

pub fn new() -> Docker

constructs a new Docker instance for a docker host listening at a url specified by an env var DOCKER_HOST, falling back on unix:///var/run/docker.sock

pub fn host(host: Url) -> Docker

constructs a new Docker instance for docker host listening at the given host url

pub fn images<'a>(&'a self) -> Images

Exports an interface for interacting with docker images

pub fn containers<'a>(&'a self) -> Containers

Exports an interface for interacting with docker containers

pub fn networks<'a>(&'a self) -> Networks

pub fn version(&self) -> Result<Version>

Returns version information associated with the docker daemon

pub fn info(&self) -> Result<Info>

Returns information associated with the docker daemon

pub fn ping(&self) -> Result<String>

Returns a simple ping response indicating the docker daemon is accessible

pub fn events(
&self,
opts: &EventsOptions
) -> Result<Box<Iterator<Item = Event>>>

Returns an interator over streamed docker events