Struct shiplift::Docker

pub struct Docker { /* fields omitted */ }

Entrypoint interface for communicating with docker daemon

Implementations

impl Docker

pub fn new() -> Docker

constructs a new Docker instance for a docker host listening at a url specified by an env var DOCKER_HOST, falling back on unix:///var/run/docker.sock

pub fn unix<S>(socket_path: S) -> Docker where
S: Into<String>,

Creates a new docker instance for a docker host listening on a given Unix socket.

pub fn host(host: Uri) -> Docker

constructs a new Docker instance for docker host listening at the given host url

pub fn images(&self) -> Images<'_>

Exports an interface for interacting with docker images

pub fn containers(&self) -> Containers<'_>

Exports an interface for interacting with docker containers

pub fn services(&self) -> Services<'_>

Exports an interface for interacting with docker services

pub fn networks(&self) -> Networks<'_>

pub fn volumes(&self) -> Volumes<'_>

pub async fn version(&self) -> Result<Version>

Returns version information associated with the docker daemon

pub async fn info(&self) -> Result<Info>

Returns information associated with the docker daemon

pub async fn ping(&self) -> Result<String>

Returns a simple ping response indicating the docker daemon is accessible

pub fn events<'docker>(
&'docker self,
opts: &EventsOptions
) -> impl Stream<Item = Result<Event>> + Unpin + 'docker

Returns a stream of docker events