Struct shiplift::Images

pub struct Images<'a> { /* fields omitted */ }

Interface for docker images

Methods

impl<'a> Images<'a>

pub fn new(docker: &'a Docker) -> Images<'a>

Exports an interface for interacting with docker images

pub fn build(&self, opts: &BuildOptions) -> Result<Box<Iterator<Item = Json>>>

Builds a new image build by reading a Dockerfile in a target directory

pub fn list(&self, opts: &ImageListOptions) -> Result<Vec<ImageRep>>

Lists the docker images on the current docker host

pub fn get(&'a self, name: &'a str) -> Image

Returns a reference to a set of operations available for a named image

pub fn search(&self, term: &str) -> Result<Vec<SearchResult>>

Search for docker images by term

pub fn pull(&self, opts: &PullOptions) -> Result<Box<Iterator<Item = Json>>>

Pull and create a new docker images from an existing image

pub fn export(&self, names: Vec<&str>) -> Result<Box<Read>>

exports a collection of named images, either by name, name:tag, or image id, into a tarball